+++ lib/core-net/service.c
@@ -421,7 +421,7 @@ lws_buflist_aware_finished_consuming(struct lws *wsi, struct lws_tokens *ebuf,
 		return 0;
 
 	if (used && buffered) {
-		m = lws_buflist_use_segment(&wsi->buflist, used);
+		m = (int)lws_buflist_use_segment(&wsi->buflist, (size_t)used);
 		// lwsl_notice("%s: used %d, next %d\n", __func__, used, m);
 		// lws_buflist_describe(&wsi->buflist, wsi, __func__);
 		if (m)
